[FIXME] ClientConnections cannot be closed on a different thread
This patch makes `ClientConnections` inside `RemoteEntityProvider` leak instead of
deleting them using `unique_ptr`s. This is a temporary fix for multi-threaded users
crashing because of a different thread trying to close client connections then the
one that created them.

In the future a better mechanism for closing the client connections upon thread exit
needs to be investigated.